Add a GitHub workflow to build installers
It is really convenient to be able to use GitHub Actions to run
automated tasks, such as creating the Git for macOS installer.

This adds a GitHub workflow definition to do exactly that.

For now, there is unfortunately no way to trigger on releases by
upstream Git, but at least we can trigger it manually, by having the
GitHub workflow listen for `repository_dispatch` event and then sending
one via:

	curl -i -H 'Authorization: token <PAT>' \
		-H 'Accept: application/vnd.github.everest-preview+json' \
		--request POST \
		--data '{
			"event_type": "osx-installer",
			"client_payload": {
				"repository": "https://github.com/git/git",
				"ref": "refs/tags/v2.26.1"
			}
		}' https://api.github.com/repos/timcharper/git_osx_installer/dispatches

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>